src/Garage.ChatService/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
openai>=1.60.0
openfeature-sdk>=0.8.2
openfeature-provider-ofrep>=0.2.0
pyyaml>=6.0.0
opentelemetry-api>=1.29.0
opentelemetry-sdk>=1.29.0
opentelemetry-instrumentation-fastapi>=0.50b0
opentelemetry-exporter-otlp-proto-http>=1.29.0
